Add public API
